project/server/pyproject.toml
[project]
name = "cua-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cua-agent[all]>=0.4.31",
    "cua-computer-server>=0.1.22",
    "openai>=1.107.2",
    "quart>=0.20.0",
    "quart-cors>=0.7.0",
]
